Room modify

+++ lib/src/main/java/com/versatica/mediasoup/MediasoupOptions.kt
@@ -22,7 +22,7 @@ class RoomOptions {
     //roomSettings
     var roomSettings: QueryRoomResponse? = null
     //Timeout for mediasoup protocol sent requests (in milliseconds)
-    var requestTimeout: Int = 30000
+    var requestTimeout: Long = 30000
     //Options for created transports.
     var transportOptions: TransportOptions = TransportOptions()
     //Array of TURN servers.
